Read bible mode #149 (#401)
* Read bible mode #149 Added new route

* added the study and read link on landing page

* Read mode #149 change landing page cards to icons

* #149 read mode merg conflict with commentary intro

* #149 Read mode , link url to bible reference & ver

* #149 read mode link url to book

* #149 Changed landing page UI and links with Samuel

* added  book code and chapter for next and previous  #149

* #149 read mode refactor code, link url

---------
