+++ package.json
@@ -22,7 +22,7 @@
     "react-redux": "^7.2.4",
     "react-redux-firebase": "^3.10.0",
     "react-router": "^5.2.0",
-    "react-router-dom": "^5.2.0",
+    "react-router-dom": "^6.14.2",
     "react-scripts": "^4.0.3",
     "react-select": "^4.3.1",
     "react-slideshow-image": "^3.5.0",
